chore: Reorganize the blocking reader layout
